% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspe_function.R
\name{mspe}
\alias{mspe}
\title{Computation of Mean Square Prediction Error (MSPE)}
\usage{
mspe(Response, Predictors)
}
\arguments{
\item{Response}{A \code{data-frame} contains the response value.}

\item{Predictors}{A \code{data-frame} contains the predictor values.}
}
\value{
A \code{data frame} containing the following attributes:
\describe{
     \item{mse}{Mean Square Prediction Error (MSPE)}
     }
}
\description{
This function computes the Mean Square Prediction Error (MSPE) in matrix form.
}
\examples{
coeff(Response, Predictors)
}
\author{
Ozan Turkes, Ayomide Afolabi, Geeta Kharel
}
